+++ src/TradeWindsDateTime/DateTimeExtensions.cs
@@ -19,24 +19,32 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using TradeWindsDateTime;
+namespace TradeWindsDateTime;
 
-namespace UnitTests
+/// <summary>
+/// Extensions to DateTime.
+/// </summary>
+public static class DateTimeExtensions
 {
+  /// <summary>
+  /// Returns the DateDiff between two dates. Does an Abs(diff) so the result is always positive.
+  /// </summary>
+  /// <param name="date1">This DateTime.</param>
+  /// <param name="date2">The comparison DateTime</param>
+  /// <returns>The difference.</returns>
+  public static DateTimeSpan DateDiff(this DateTime date1, DateTime date2)
+  {
+    return DateTimeSpan.Diff(date1, date2);
+  }
 
-	public class TestDateTimeExtensions
-	{
-		[Fact]
-		public void TestAll()
-		{
-
-			var dateTime = new DateTime(1955, 9, 26, 1, 2, 29);
-			var dateTimeRound = dateTime.RoundToMinute();
-			Assert.Equal(new DateTime(1955, 9, 26, 1, 2, 0), dateTimeRound);
-
-			var dateTime2 = new DateTime(1955, 9, 26, 1, 2, 31);
-			var dateTimeRound2 = dateTime2.RoundToMinute();
-			Assert.Equal(new DateTime(1955, 9, 26, 1, 3, 0), dateTimeRound2);
-		}
-	}
+  /// <summary>
+  /// Round to the nearest minute.
+  /// </summary>
+  /// <param name="dt">The DateTime to round.</param>
+  /// <returns>The rounded value.</returns>
+  public static DateTime RoundToMinute(this DateTime dt)
+  {
+    var truncated = new DateTime(dt.Year, dt.Month, dt.Day, dt.Hour, dt.Minute, 0, dt.Kind);
+    return dt.Second >= 30 ? truncated.AddMinutes(1) : truncated;
+  }
 }
