+++ src/TradeWindsDateTime/TradeWindsDateTime.csproj
@@ -4,7 +4,7 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-	<Version>1.0.0.1</Version>
+    <Version>1.0.0.1</Version>
   </PropertyGroup>
 
 </Project>
