+++ tests/UnitTests/TestDateTimeConvertor.cs
@@ -19,33 +19,27 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-namespace TradeWindsDateTime
-{
-	/// <summary>
-	/// Extensions to DateTime.
-	/// </summary>
-	public static class DateTimeExtensions
-	{
-		/// <summary>
-		/// Returns the DateDiff between two dates. Does an Abs(diff) so the result is always positive.
-		/// </summary>
-		/// <param name="date1">This DateTime.</param>
-		/// <param name="date2">The comparison DateTime</param>
-		/// <returns>The difference.</returns>
-		public static DateTimeSpan DateDiff(this DateTime date1, DateTime date2)
-		{
-			return DateTimeSpan.Diff(date1, date2);
-		}
+using TradeWindsDateTime;
 
-		/// <summary>
-		/// Round to the nearest minute.
-		/// </summary>
-		/// <param name="dt">The DateTime to round.</param>
-		/// <returns>The rounded value.</returns>
-		public static DateTime RoundToMinute(this DateTime dt)
-		{
-			var truncated = new DateTime(dt.Year, dt.Month, dt.Day, dt.Hour, dt.Minute, 0, dt.Kind);
-			return dt.Second >= 30 ? truncated.AddMinutes(1) : truncated;
-		}
-	}
+namespace UnitTests;
+
+public class TestDateTimeConvertor
+{
+  [Fact]
+  public void TestAll()
+  {
+
+    var convertor = DateTimeConvertor.Create("Mountain Standard Time");
+
+    var dateTime = new DateTime(1955, 9, 26, 1, 2, 3);
+    var dateTimeUtc = convertor.ConvertToUtc(dateTime);
+    Assert.Equal(dateTime.Date, dateTimeUtc.Date);
+    // the hour is off by 6 or 7 because of the timezone (7 is DST)
+    Assert.True(dateTimeUtc.Hour == dateTime.Hour + 6 || dateTimeUtc.Hour == dateTime.Hour + 7);
+    Assert.Equal(dateTime.Minute, dateTimeUtc.Minute);
+    Assert.Equal(dateTime.Second, dateTimeUtc.Second);
+
+    var dateTimeLocal = convertor.ConvertFromUtc(dateTimeUtc);
+    Assert.Equal(dateTime, dateTimeLocal);
+  }
 }
