+++ tests/UnitTests/UnitTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -20,10 +20,11 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
+
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TradeWindsDateTime\TradeWindsDateTime.csproj" />
+    <ProjectReference Include="..\..\src\TradeWindsDateTime\TradeWindsDateTime.csproj" />
   </ItemGroup>
 
   <ItemGroup>
